Extract input handling and pipeline steps from index entrypoint
